2021/10/19 Fifth commit

+++ app/src/main/java/com/example/risingproject/src/main/store/storehome/models/ResultItemUseFilter.kt
@@ -12,5 +12,5 @@ data class ResultItemUseFilter (
     @SerializedName("percenttage") val percenttage: Int,
     @SerializedName("deliveryFee") val deliveryFee: String,
     @SerializedName("numOfReviews") val numOfReviews: Int?,
-    @SerializedName("reviewRate") val reviewRate: Int
+    @SerializedName("reviewRate") val reviewRate: Double
     )
